First pass at attempting to port Makefile to grunt
I don't expect this to be merged yet. I want to get feedback.

The Makefile is really hard to understand (for me at least), so I'm
taking a stab at porting it to grunt. This is a rough first iteration,
and doesn't result result in a working build (yet). It shows that we
could accomplish many of the things we use make for, and (hopefully) in
a cleaner fashion. It also uses various third-party grunt modules to
replace the functionality of many helper scripts, like
`mustache_render.js`.

For this first iteration, I tried not to modify any of the source tree,
but in the long run, this could lay the foundation for further
improvements, like downloading most/all dependencies from npm or bower.

Let me know what you think works, and what doesn't. If you think this is
the wrong direction and feel that SCons is the javascript build system
of the future, let me know that too.

I don't think Make is evil/boring/uncool. Quite the opposite
<https://github.com/PiPeep/prelude.mk>. I just feel like this project
has long-since outgrown it.